+++ front/src/components/common/TopNav.jsx
@@ -1,17 +1,13 @@
 import { BellOutlined, MailOutlined, GiftOutlined } from "@ant-design/icons";
-import {
-  Wrapper,
-  BackIcon,
-  Title,
-  IconGroup,
-  HeaderIcon,
-} from "./TopNav.styles";
+import { Wrapper, BackIcon, Title, IconGroup, HeaderIcon } from "./TopNav.styles";
 import { Badge } from "antd";
-
+const wsNotificationIp = import.meta.env.VITE_WS_NOTIFICATION_IP || "";
 import UserDropDown from "../../components/common/UserDropDwonMenu";
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
 import { fetchUnreadCount } from "../../store/thunks/notificationThunks";
+import { fetchMyInfo } from "../../store/thunks/authThunks";
+import { increaseUnreadCount } from "../../store/slices/notificationSlice";
 
 export default function TopNav({ title, onBack, onNotification, onMessage }) {
   const dispatch = useDispatch();
@@ -20,6 +16,40 @@ export default function TopNav({ title, onBack, onNotification, onMessage }) {
   useEffect(() => {
     dispatch(fetchUnreadCount());
   }, [dispatch]);
+
+  const userId = useSelector((state) => state.auth.userId);
+
+  const protocol = window.location.protocol === "https:" ? "wss" : "ws";
+
+  useEffect(() => {
+    dispatch(fetchMyInfo());
+  }, [dispatch]);
+
+  useEffect(() => {
+    if (userId == null) return;
+    const socket = new WebSocket(`${protocol}${wsNotificationIp}`);
+    socket.onopen = () => {
+      console.log("Notification WS connected");
+    };
+    socket.onmessage = (event) => {
+      console.log("Notification WS message received");
+      const msg = event.data;
+      console.log("Notification WS message content:", msg);
+      if (msg === "Notification") {
+        dispatch(increaseUnreadCount());
+      }
+    };
+
+    socket.onerror = (e) => {
+      console.error("WebSocket error", e);
+    };
+    socket.onclose = () => {
+      console.log("Notification WS disconnected");
+    };
+    return () => {
+      socket.close();
+    };
+  }, [dispatch, userId]);
 
   return (
     <Wrapper>
@@ -35,12 +65,7 @@ export default function TopNav({ title, onBack, onNotification, onMessage }) {
         {onNotification && (
           <HeaderIcon onClick={onNotification}>
             <BellOutlined style={{ fontSize: "3.5vh", color: "inherit" }} />
-            <Badge
-              count={unreadCount > 0 ? unreadCount : null}
-              size="small"
-              overflowCount={99}
-              offset={[0, 4]}
-            ></Badge>
+            <Badge count={unreadCount > 0 ? unreadCount : null} size="small" overflowCount={99} offset={[0, 4]}></Badge>
           </HeaderIcon>
         )}
